requirements.txt
# Core MCP e LangChain
mcp
mcp[cli]
langchain
langchain-groq
langchain-anthropic
langchain-openai
langchain-mcp-adapters
langgraph

# Utilities
python-dotenv
loguru

# Chainlit
chainlit